Fix label flickering
- Fixes bug where labels flickered randomly when object was aligned with
the camera direction vector. We moved the distToCamera attribute from
float to dobule to reduce the angle determination error. This fixes #122

+++ core/src/gaia/cu9/ari/gaiaorbit/render/IRenderable.java
@@ -19,6 +19,6 @@ public interface IRenderable {
      * Gets the last distance to the camera calculated for this entity.
      * @return The distance.
      */
-    public float getDistToCamera();
+    public double getDistToCamera();
 
 }
