+++ core/src/gaia/cu9/ari/gaiaorbit/util/comp/DistToCameraComparator.java
@@ -1,8 +1,8 @@
 package gaia.cu9.ari.gaiaorbit.util.comp;
 
-import gaia.cu9.ari.gaiaorbit.scenegraph.AbstractPositionEntity;
-
 import java.util.Comparator;
+
+import gaia.cu9.ari.gaiaorbit.scenegraph.AbstractPositionEntity;
 
 /**
  * Compares entities. Further entities go first, nearer entities go last.
@@ -14,7 +14,7 @@ public class DistToCameraComparator<T> implements Comparator<T> {
 
     @Override
     public int compare(T o1, T o2) {
-        return -Float.compare(((AbstractPositionEntity) o1).distToCamera, ((AbstractPositionEntity) o2).distToCamera);
+        return -Double.compare(((AbstractPositionEntity) o1).distToCamera, ((AbstractPositionEntity) o2).distToCamera);
     }
 
 }
